requirements.txt
duckduckgo_search
mistralai
Pillow
pydantic
pyprojroot
python-dotenv
PyYAML
streamlit
streamlit_chat